INTEG-344 [integ-forum] NPE in Poll activity of deleted space
Fix description: In PollUIActivity template, if uicomponent.getVotes() returns null, then we know that the poll is deleted. In this case, do not continue to render groovy template. Use method isDeletedSpace from Social project to check the space is deleted, and in this case, do not get vote, nor render html code.