INTEG-486 Avoid display document name (#164)
The activity title have a default value when it's empty, which is uploaded file names. This was made before 5.0 and thus to change this, even if it's not used, we will have to make an upgrade plugin to make the activity title holds only user message (which is complex and useless for now). So this will simply verify if the original written message from user (without pre and post save transformation) is empty. If it's empty, it means the original user post was empty, else, we should display the processed title coming from user post.