ECMS-7863 : use system session provider to be able to get user node (#207)
The session used to get the folder where the activity's file will be stored was an user session, which did not allowed to get it.
Indeed, in order to get the personal user node, the NodeHierarchyCreator service uses the DataDistributionType service which needs to get the Users root node first.
But this node is only accessible by administrators (on purpose). So since we use an user session, it is forbidden.

This fix uses a system session to get the user node folder and stores the file.