SOC-6193 : replace embedded images link in activity template params (#211)
When add an embedded image in an activity post (with CKEditor), CKEditor uses a temporary link in the HTML to upload it.
When the activity is posted, this temporary link is replaced by the final link. This replacement is done in the title and body attributes of an activity.
But the activities of type Link, for some reasons, do not store the activity post content in the title or body attribute, but in a template params.
Since the links in template params are not replaced, the temporary link is kept, which leads to lost the image when the link is not available anymore (for example when the server is restarted).

This fix replace the CKEditor temporary images links also in template params to make sure they are replaced everywhere.