CAL-1443 : do not rely on WebUI to get locale to use to generate activity comment (#212)
When an event is updated, the activity related to this event is commented to described the changes.
The body of the comment is generated according to the user locale. This locale was retrieved from WebUI context.
Since the implementation of Create Event improvement, the events are updated through REST services and the WebUI context is not available anymore, thats' why is comment generation failed.

This fix does not rely anymore on the WebUI context in the comment generation. Instead the Locale is injected from the template.