CAL-1438 : use same method everychere to check if an event is allDay or not (#214)
When updating the activity linked to an event, we check if the event is on all day or not in order to display it in the activity.
In that case the check is not done like in others cases. It uses full timestamp instead of only day/hours/minutes.

This fix changes the way the check on allDay is done in this case to align with others cases.