PLF-8383 : refactor News details to make code more reusable (#243)
This change creates a new component which displays the whole News details (including edit and share features), in order to reuse this components in several locations : in the activity detail page and in content view template.
This commit goes a bit further by extracting generic components as reusable components : modal, suggester and fileDrop. These components are now defined as AMD modules and can be injected in any other script/component.